+++ src/components/General/CustomControls.tsx
@@ -2,7 +2,7 @@ import { createPortal } from "react-dom";
 import { useEffect, useRef, forwardRef } from "react";
 import { createControlComponent } from "@react-leaflet/core";
 import { Control } from "leaflet";
-import { createControlContainer } from "../../services";
+import { createControlContainer } from "../../services/cards";
 import { ControlProps, ChildrenProp } from "../../types";
 
 const ControlContainer = createControlComponent<Control, ControlProps>(
