+++ src/components/Introduction/components/IntroNav.tsx
@@ -2,7 +2,7 @@ import { RefObject } from "react";
 
 import { AboutHeader, GeneralLink } from "../../styled_components";
 
-import { intro } from "../../../services";
+import { intro } from "../../../services/navigation";
 
 function IntroNav(props: {
   howToRef: RefObject<HTMLHeadingElement>;
