+++ src/services/divIconFactory.ts
